# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Harness daemon API",
    "description": "REST API served by the Harness daemon (harness-daemon). The API is dual-transport: most endpoints mirror a WebSocket JSON-RPC method, recorded per operation as the `x-websocket-method` extension. Generated from annotated handlers - do not edit docs/api/openapi.json by hand; regenerate with `mise run openapi:generate`.",
    "license": {
      "name": "MIT"
    },
    "version": "50.0.0"
  },
  "paths": {
...
    "/v1/sessions/{session_id}/agents/{session_agent_id}/remove": {
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "post_remove_agent",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "session_agent_id",
            "in": "path",
            "description": "Session-scoped agent identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRemoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent removed from the session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "agent.remove"
      }
    },
    "/v1/sessions/{session_id}/agents/{session_agent_id}/role": {
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "post_role_change",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "session_agent_id",
            "in": "path",
            "description": "Session-scoped agent identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent role changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "agent.change_role"
      }
    },
...
    "/v1/sessions/{session_id}/leader": {
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "post_transfer_leader",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaderTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session leadership transferred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Request error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DaemonErrorBody"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/RemoteAuthRequired"
          },
          "414": {
            "$ref": "#/components/responses/RemoteRequestUriTooLong"
          },
          "413": {
            "$ref": "#/components/responses/RemoteRequestBodyTooLarge"
          },
          "431": {
            "$ref": "#/components/responses/RemoteRequestHeadersTooLarge"
          },
          "429": {
            "$ref": "#/components/responses/RemoteRequestThrottled"
          },
          "503": {
            "$ref": "#/components/responses/RemoteServiceUnavailable"
          },
          "504": {
            "$ref": "#/components/responses/RemoteRequestTimedOut"
          }
        },
        "x-websocket-method": "leader.transfer"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentPendingUserPrompt": {
        "type": "object",
        "required": [
          "tool_name"
        ],
        "properties": {
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Compatibility summary for clients that still expect a single-line prompt\nmessage."
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AskUserQuestionPrompt"
            }
          },
          "tool_name": {
            "type": "string"
          },
          "waiting_since": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "AgentPersona": {
        "type": "object",
        "description": "A predefined agent definition that shapes an agent's role and focus.",
        "required": [
          "identifier",
          "name",
          "symbol",
          "description"
        ],
        "properties": {
          "description": {
            "type": "string",
            "description": "What this persona does, shown in detail views."
          },
          "identifier": {
            "type": "string",
            "description": "Unique slug (e.g. `code-reviewer`, `test-writer`)."
          },
          "name": {
            "type": "string",
            "description": "Human-readable display name."
          },
          "symbol": {
            "$ref": "#/components/schemas/PersonaSymbol",
            "description": "Icon for visual identification."
          }
        }
      },
      "AgentRegistrationWire": {
        "type": "object",
        "required": [
          "session_agent_id",
          "name",
          "runtime",
          "role",
          "joined_at",
          "updated_at",
          "status"
        ],
        "properties": {
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "current_task_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "descriptor_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "joined_at": {
            "type": "string"
          },
          "last_activity_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "managed_agent_family": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ManagedAgentKind"
              }
            ]
          },
          "managed_agent_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "persona": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentPersona"
              }
            ]
          },
          "role": {
            "$ref": "#/components/schemas/SessionRole"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeKindSchema"
          },
          "runtime_capabilities": {
            "$ref": "#/components/schemas/RuntimeCapabilities"
          },
          "runtime_session_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "runtime_session_title": {
            "type": [
              "string",
              "null"
            ]
          },
          "session_agent_id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AgentStatusSchema"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
      "AgentRemoveRequest": {
        "type": "object",
        "required": [
          "actor"
        ],
        "properties": {
          "actor": {
            "type": "string"
          }
        }
      },
...
      "AgentToolActivitySummary": {
        "type": "object",
        "required": [
          "agent_id",
          "runtime",
          "tool_invocation_count",
          "tool_result_count",
          "tool_error_count",
          "recent_tools"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "latest_event_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "latest_tool_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "pending_user_prompt": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AgentPendingUserPrompt"
              }
            ]
          },
          "recent_tools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "runtime": {
            "type": "string"
          },
          "tool_error_count": {
            "type": "integer",
            "minimum": 0
          },
          "tool_invocation_count": {
            "type": "integer",
            "minimum": 0
          },
          "tool_result_count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "DaemonErrorBody": {
        "type": "object",
        "description": "Error envelope returned by daemon handlers on failure.\n\nMirrors the dominant shape produced by `error_status_and_body`\n(`{\"error\": {\"code\", \"message\", \"details\"}}`). A few endpoints emit\nalternate ad-hoc error shapes for specific conditions; those are noted in\nthe relevant operation descriptions.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/DaemonErrorDetail"
          }
        }
      },
      "DaemonErrorDetail": {
        "type": "object",
        "description": "Structured error detail carried by [`DaemonErrorBody`].",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code (for example `SESSION_SCOPE_DENIED`)."
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional additional context lines."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "LeaderTransferRequest": {
        "type": "object",
        "required": [
          "actor",
          "new_leader_id"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "new_leader_id": {
            "type": "string"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ManagedAgentKind": {
        "type": "string",
        "description": "Transport family for a daemon-managed agent reference.",
        "enum": [
          "tui",
          "acp",
          "codex"
        ]
      },
...
      "ObserverSummary": {
        "type": "object",
        "required": [
          "observe_id",
          "last_scan_time",
          "open_issue_count",
          "resolved_issue_count",
          "muted_code_count",
          "active_worker_count",
          "open_issues",
          "muted_codes",
          "active_workers",
          "agent_sessions"
        ],
        "properties": {
          "active_worker_count": {
            "type": "integer",
            "minimum": 0
          },
          "active_workers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObserverActiveWorker"
            }
          },
          "agent_sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObserverAgentSessionSummary"
            }
          },
          "last_scan_time": {
            "type": "string"
          },
          "muted_code_count": {
            "type": "integer",
            "minimum": 0
          },
          "muted_codes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IssueCode"
            }
          },
          "observe_id": {
            "type": "string"
          },
          "open_issue_count": {
            "type": "integer",
            "minimum": 0
          },
          "open_issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObserverOpenIssue"
            }
          },
          "resolved_issue_count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "RoleChangeRequest": {
        "type": "object",
        "required": [
          "actor",
          "role"
        ],
        "properties": {
          "actor": {
            "type": "string"
          },
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/SessionRole"
          }
        }
      },
...
      "RuntimeKindSchema": {
        "type": "object",
        "required": [
          "kind",
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Runtime identifier within the family."
          },
          "kind": {
            "type": "string",
            "description": "Runtime family, `tui` or `acp`."
          }
        }
      },
...
      "SessionDetail": {
        "type": "object",
        "required": [
          "session",
          "agents",
          "tasks",
          "signals",
          "agent_activity"
        ],
        "properties": {
          "agent_activity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentToolActivitySummary"
            }
          },
          "agents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentRegistrationWire"
            }
          },
          "observer": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ObserverSummary"
              }
            ]
          },
          "session": {
            "$ref": "#/components/schemas/SessionSummary"
          },
          "signals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionSignalRecord"
            }
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkItem"
            }
          }
        }
      },
...
      "SessionRole": {
        "type": "string",
        "description": "Role an agent holds within a session.",
        "enum": [
          "leader",
          "observer",
          "worker",
          "reviewer",
          "improver"
        ]
      },
      "SessionSignalRecord": {
        "type": "object",
        "description": "A signal visible within a multi-agent session.",
        "required": [
          "runtime",
          "agent_id",
          "session_id",
          "status",
          "signal"
        ],
        "properties": {
          "acknowledgment": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SignalAck"
              }
            ]
          },
          "agent_id": {
            "type": "string"
          },
          "runtime": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "signal": {
            "$ref": "#/components/schemas/Signal"
          },
          "status": {
            "$ref": "#/components/schemas/SessionSignalStatus"
          }
        }
      },
...
      "SessionStatus": {
        "type": "string",
        "description": "Session lifecycle status.",
        "enum": [
          "awaiting_leader",
          "active",
          "paused",
          "leaderless_degraded",
          "ended"
        ]
      },
      "SessionSummary": {
        "type": "object",
        "required": [
          "project_id",
          "project_name",
          "context_root",
          "worktree_path",
          "shared_path",
          "origin_path",
          "branch_ref",
          "session_id",
          "title",
          "context",
          "status",
          "created_at",
          "updated_at",
          "metrics"
        ],
        "properties": {
          "adopted_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "branch_ref": {
            "type": "string"
          },
          "context": {
            "type": "string"
          },
          "context_root": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "external_origin": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_activity_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "leader_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "metrics": {
            "$ref": "#/components/schemas/SessionMetrics"
          },
          "observe_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "origin_path": {
            "type": "string"
          },
          "pending_leader_transfer": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PendingLeaderTransfer"
              }
            ]
          },
          "project_dir": {
            "type": [
              "string",
              "null"
            ]
          },
          "project_id": {
            "type": "string"
          },
          "project_name": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "shared_path": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "worktree_path": {
            "type": "string"
          }
        }
      },
...
      "WorkItem": {
        "type": "object",
        "description": "A work item tracked within a session.",
        "required": [
          "task_id",
          "title",
          "severity",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "arbitration": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ArbitrationOutcome",
                "description": "Leader arbitration outcome once recorded."
              }
            ]
          },
          "assigned_to": {
            "type": [
              "string",
              "null"
            ]
          },
          "awaiting_review": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AwaitingReview",
                "description": "Metadata for the awaiting-review queue entry."
              }
            ]
          },
          "blocked_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "checkpoint_summary": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TaskCheckpointSummary"
              }
            ]
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "consensus": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReviewConsensus",
                "description": "Closed quorum consensus, set once `required_consensus` reviewers agree."
              }
            ]
          },
          "context": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "created_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "deleted_at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Tombstone timestamp for tasks deleted from active session views while\npreserving their historical identity and timeline."
          },
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskNote"
            }
          },
          "observe_issue_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "queue_policy": {
            "$ref": "#/components/schemas/TaskQueuePolicy"
          },
          "queued_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "review_claim": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReviewClaim",
                "description": "Active reviewer claim, including distinct-runtime reviewer entries."
              }
            ]
          },
          "review_history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewConsensus"
            },
            "description": "Per-round consensus history: each entry is the merged-and-responded\nconsensus from one completed round, pushed here before\n`task.consensus` is cleared so per-point agree/dispute state and\nworker notes survive across rounds."
          },
          "review_round": {
            "type": "integer",
            "format": "int32",
            "description": "Review round counter; increments each time a reworked task goes back to review.",
            "minimum": 0
          },
          "severity": {
            "$ref": "#/components/schemas/TaskSeverity"
          },
          "source": {
            "$ref": "#/components/schemas/TaskSource"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "suggested_fix": {
            "type": [
              "string",
              "null"
            ]
          },
          "suggested_persona": {
            "type": [
              "string",
              "null"
            ],
            "description": "Persona hint to bias routing when assigning to a worker."
          },
          "task_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
      "RemoteAuthRequired": {
        "description": "Missing, invalid, or unauthorized daemon credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestUriTooLong": {
        "description": "Request URI exceeds the configured remote limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestBodyTooLarge": {
        "description": "Request body exceeds the configured remote limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestHeadersTooLarge": {
        "description": "Request headers exceed the configured remote limit",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestThrottled": {
        "description": "Remote request or connection concurrency limit reached",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteServiceUnavailable": {
        "description": "Remote auth store or request limits unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      },
      "RemoteRequestTimedOut": {
        "description": "Remote request exceeded the configured timeout",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DaemonErrorBody"
            }
          }
        }
      }
    }
  },
...
}